backend/requirements.txt
Flask==2.1.2
flask_pymongo==2.3.0
pymongo==4.0.2
PyJWT==2.3.0
flask_cors==3.0.10
pytz==2022.1
python-dotenv
